+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 - Identifier schema is not a Registry, it's a regular entity
 - Add default: `now` for datetime fields
+- Add Category kinds and default kind
 
 ## [1.0.3][] - 2021-06-30